[FIX] constData 파일 -> data 폴더로 분화
