+++ src/actions/commentActions.ts
@@ -1,8 +1,8 @@
 "use server";
 
-import { tempAccessToken } from "@/assets/tempAccessToken";
 import { AUTHORIZATION_HEADERS, COMMON_HEADERS } from "@/config/httpRequestHeaders";
 import { COMMENT_API_URL } from "@/constants/apiEndpoints";
+import { tempAccessToken } from "@/data/tempAccessToken";
 import type { Comment } from "@/types/commentTypes";
 import type { MainCategory } from "@/types/filterTypes";
 import { handleError } from "@/utils/errorHandler";
