+++ src/data/categoryData.ts
@@ -1,13 +1,3 @@
-import {
-  Chat,
-  FaceNeutral,
-  FaceSmile,
-  LineHeart,
-  MessageCircle,
-  MessageSquare,
-  MyUser,
-  Users,
-} from "@/assets/svgs/icons";
 import type { MainCategory } from "@/types/filterTypes";
 
 export const MAIN_CATEGORY: MainCategory[] = ["PJ", "PR"];
@@ -63,18 +53,4 @@ export const FILTER_MENU = [
   },
 ];
 
-export const NAV_MENU = [
-  { title: "모임", Component: Users, route: "/" },
-  { title: "채팅", Component: Chat, route: "/chat" },
-  { title: "활동", Component: LineHeart, route: "/dashboard" },
-  { title: "마이", Component: MyUser, route: "/mypage" },
-];
-
 export const DUTY_CATEGORY = ["게임", "기획", "디자인", "모바일/APP", "백엔드", "프론트엔드"];
-
-export const NOTIFICATION_TYPE = {
-  accept: { icon: FaceSmile, message: (subject: string) => `\`${subject}\`에 수락됐어요.` },
-  reject: { icon: FaceNeutral, message: (subject: string) => `\`${subject}\`에 거절됐어요.` },
-  comment: { icon: MessageSquare, message: (subject: string) => `\`${subject}\`에 새로운 댓글이 달렸어요.` },
-  chat: { icon: MessageCircle, message: (subject: string) => `\`${subject}\`님과 새로운 채팅이 시작됐어요.` },
-};
